refactor: вынес компонент PlayersListTable отдельно от PlayersList
