+++ packages/client/src/components/PlayersList/index.tsx
@@ -1,28 +1,11 @@
-import {Avatar} from 'client/src/components/Avatar';
-import React, {FC, Fragment} from 'react';
+import React, {FC} from 'react';
 import styles from './playersList.module.scss';
 import {TPlayersListProps} from './typing';
+import {PlayersListTable} from './Table';
 
 export const PlayersList: FC<TPlayersListProps> = React.memo((props) => {
 
 	const {data, showIndex = false} = props;
-
-	const PlayersListTable = () => {
-		const lobbyDataItems = data.map((playerData, index) => {
-			return (
-				<Fragment>
-					<div data-column='index'>{(index)}</div>
-					<div data-column='nickname'>
-						<Avatar alt='player' size='small'/>
-						{playerData.nickname}
-					</div>
-					<div>{playerData.winsOnline}</div>
-					<div>{playerData.winsOffline}</div>
-				</Fragment>
-			);
-		});
-		return <>{lobbyDataItems}</>;
-	};
 
 	return (
 		<div className={styles.wrapper}>
@@ -32,7 +15,7 @@ export const PlayersList: FC<TPlayersListProps> = React.memo((props) => {
 				<div>Игрок</div>
 				<div>Побед онлайн</div>
 				<div>Побед оффлайн</div>
-				<PlayersListTable/>
+				<PlayersListTable data={data}/>
 			</div>
 		</div>
 	);
